ic-score-service/requirements.txt
# InvestorCenter.ai IC Score Service - Python Dependencies
# Python 3.11+ required

# SQLAlchemy with async support
sqlalchemy[asyncio]>=2.0.0

# Database drivers
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.9  # PostgreSQL sync driver (for Alembic)

# Database migrations
alembic>=1.12.0

# Date/time utilities
python-dateutil>=2.8.2

# Environment variables
python-dotenv>=1.0.0

# API clients (Phase 1-4)
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Data processing (Phase 1-2)
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0  # For risk metrics calculations

# Technical analysis (Phase 2)
# pandas-ta>=0.3.14b0  # REMOVED: Not available for Python 3.11/ARM64, using TA-Lib instead
TA-Lib>=0.4.28  # Note: requires system-level installation

# NLP/Sentiment (Phase 4)
transformers>=4.35.0
torch>=2.1.0
sentencepiece>=0.1.99

# FastAPI (Phase 5)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
tqdm>=4.66.0
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0
